working on coin_prices.py
