+++ tracker/db/price.py
@@ -24,6 +24,7 @@ def select(date_: Optional[Union[str, date]] = None) -> List[dict]:
 
 
 def get_missing() -> List[dict]:
+    """ find which data are missing from the prices table, based on date and coin present in coin_balances """
     conn = get_conn()
     curr = conn.cursor()
