+++ tracker/recurrent_update.py
@@ -4,7 +4,7 @@ from dateutil.relativedelta import relativedelta
 
 from tracker.coin_balances import add_missing_coin_balances
 from tracker.graph_drawer import draw_and_send
-from tracker.prices import prices_update
+from tracker.coin_prices import add_missing_coin_prices
 from tracker.tot_balances import tot_balances_update
 
 
@@ -13,7 +13,7 @@ def recurrent_update():
     Do all updates necessary (coin_balances, prices and total_balances) and the updated graph.
     """
     add_missing_coin_balances()
-    prices_update()
+    add_missing_coin_prices()
     tot_balances_update()
 
     # all
